Step3(refactor) : App 모듈에서 room dependency 제거 + 잔여 이슈 수정 + 코드 정리
