+++ calculator-data/src/main/java/edu/nextstep/camp/calculator/data/EvaluationRecordDao.kt
@@ -3,12 +3,11 @@ package edu.nextstep.camp.calculator.data
 import androidx.room.Dao
 import androidx.room.Insert
 import androidx.room.Query
-import kotlinx.coroutines.flow.Flow
 
 @Dao
 interface EvaluationRecordDao {
     @Query("SELECT * FROM evaluation_record")
-    fun getAll(): Flow<List<EvaluationRecordEntity>>
+    fun getAll(): List<EvaluationRecordEntity>
 
     @Insert
     fun insert(record: EvaluationRecordEntity)
