+++ calculator-data/src/main/java/edu/nextstep/camp/calculator/data/di/AppModule.kt
@@ -1,4 +1,4 @@
-package edu.nextstep.camp.calculator.di
+package edu.nextstep.camp.calculator.data.di
 
 import android.content.Context
 import dagger.Module
