+++ calculator-data/src/main/java/edu/nextstep/camp/calculator/data/di/RepositoryModule.kt
@@ -1,13 +1,10 @@
-package edu.nextstep.camp.calculator.di
+package edu.nextstep.camp.calculator.data.di
 
 import dagger.Binds
 import dagger.Module
-import dagger.Provides
 import dagger.hilt.InstallIn
 import dagger.hilt.android.components.ViewModelComponent
-import dagger.hilt.android.scopes.ViewModelScoped
 import edu.nextstep.camp.calculator.data.repository.EvaluationRecordRepositoryImpl
-import edu.nextstep.camp.calculator.domain.Calculator
 import edu.nextstep.camp.calculator.domain.repository.EvaluationRecordRepository
 
 @Module
@@ -16,13 +13,4 @@ abstract class RepositoryModule {
 
     @Binds
     abstract fun bindEvaluationRecordStoreRepository(repository: EvaluationRecordRepositoryImpl): EvaluationRecordRepository
-}
-
-@Module
-@InstallIn(ViewModelComponent::class)
-class DomainModule {
-
-    @Provides
-    @ViewModelScoped
-    fun provideCalculator(): Calculator = Calculator()
 }
