+++ calculator-data/src/main/java/edu/nextstep/camp/calculator/data/repository/EvaluationRecordRepositoryImpl.kt
@@ -3,10 +3,8 @@ package edu.nextstep.camp.calculator.data.repository
 import edu.nextstep.camp.calculator.data.CalculatorDatabase
 import edu.nextstep.camp.calculator.data.toEntity
 import edu.nextstep.camp.calculator.data.toEvaluationRecordList
-import edu.nextstep.camp.calculator.domain.EvaluationRecord
+import edu.nextstep.camp.calculator.domain.model.EvaluationRecord
 import edu.nextstep.camp.calculator.domain.repository.EvaluationRecordRepository
-import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.map
 import javax.inject.Inject
 
 class EvaluationRecordRepositoryImpl @Inject constructor(
@@ -16,7 +14,7 @@ class EvaluationRecordRepositoryImpl @Inject constructor(
         db.evaluationRecordDao().insert(evaluationRecord.toEntity())
     }
 
-    override suspend fun getEvaluationHistory(): Flow<List<EvaluationRecord>> {
-        return db.evaluationRecordDao().getAll().map { it.toEvaluationRecordList() }
+    override fun getEvaluationHistory(): List<EvaluationRecord> {
+        return db.evaluationRecordDao().getAll().toEvaluationRecordList()
     }
 }
