--- calculator-domain/src/main/java/edu/nextstep/camp/calculator/domain/EvaluationRecordStore.kt
@@ -1,11 +1,0 @@
-package edu.nextstep.camp.calculator.domain
-
-class EvaluationRecordStore {
-    private val evaluationHistory = mutableListOf<EvaluationRecord>()
-
-    fun record(evaluationRecord: EvaluationRecord) {
-        evaluationHistory.add(evaluationRecord)
-    }
-
-    fun getEvaluationHistory() = evaluationHistory.toList()
-}
