+++ calculator-domain/src/main/java/edu/nextstep/camp/calculator/domain/model/EvaluationRecord.kt
@@ -1,4 +1,4 @@
-package edu.nextstep.camp.calculator.domain
+package edu.nextstep.camp.calculator.domain.model
 
 data class EvaluationRecord(val expression: String, val result: String)
