+++ calculator-domain/src/main/java/edu/nextstep/camp/calculator/domain/model/Operator.kt
@@ -1,4 +1,4 @@
-package edu.nextstep.camp.calculator.domain
+package edu.nextstep.camp.calculator.domain.model
 
 enum class Operator(
     val sign: String,
