+++ calculator-domain/src/main/java/edu/nextstep/camp/calculator/domain/repository/EvaluationRecordRepository.kt
@@ -1,9 +1,8 @@
 package edu.nextstep.camp.calculator.domain.repository
 
-import edu.nextstep.camp.calculator.domain.EvaluationRecord
-import kotlinx.coroutines.flow.Flow
+import edu.nextstep.camp.calculator.domain.model.EvaluationRecord
 
 interface EvaluationRecordRepository {
     suspend fun record(evaluationRecord: EvaluationRecord)
-    suspend fun getEvaluationHistory() : Flow<List<EvaluationRecord>>
+    fun getEvaluationHistory() : List<EvaluationRecord>
 }
