+++ calculator/src/main/java/edu/nextstep/camp/calculator/CalculatorActivity.kt
@@ -8,8 +8,6 @@ import androidx.appcompat.app.AppCompatActivity
 import androidx.core.view.isVisible
 import dagger.hilt.android.AndroidEntryPoint
 import edu.nextstep.camp.calculator.databinding.ActivityCalculatorBinding
-import edu.nextstep.camp.calculator.domain.Expression
-import edu.nextstep.camp.calculator.domain.Operator
 
 @AndroidEntryPoint
 class CalculatorActivity : AppCompatActivity() {
