+++ calculator/src/main/java/edu/nextstep/camp/calculator/CalculatorViewModel.kt
@@ -5,11 +5,10 @@ import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import dagger.hilt.android.lifecycle.HiltViewModel
-import edu.nextstep.camp.calculator.data.repository.EvaluationRecordRepositoryImpl
 import edu.nextstep.camp.calculator.domain.Calculator
-import edu.nextstep.camp.calculator.domain.EvaluationRecord
-import edu.nextstep.camp.calculator.domain.Expression
-import edu.nextstep.camp.calculator.domain.Operator
+import edu.nextstep.camp.calculator.domain.model.EvaluationRecord
+import edu.nextstep.camp.calculator.domain.model.Expression
+import edu.nextstep.camp.calculator.domain.model.Operator
 import edu.nextstep.camp.calculator.domain.repository.EvaluationRecordRepository
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
@@ -19,8 +18,8 @@ import javax.inject.Inject
 @HiltViewModel
 class CalculatorViewModel @Inject constructor(
     private val evaluationRecordRepository: EvaluationRecordRepository,
-    private val calculator: Calculator
     ): ViewModel() {
+    private val calculator = Calculator()
     private var expression = Expression.EMPTY
 
     private val _state = MutableLiveData<State>(State.ShowExpression(Expression.EMPTY))
@@ -78,7 +77,9 @@ class CalculatorViewModel @Inject constructor(
         when (val currentState = _state.value) {
             is State.ShowExpression -> {
                 expression = currentState.expression
-//                _state.value = State.ShowHistory(evaluationRecordRepository.getEvaluationHistory())
+                viewModelScope.launch {
+                    _state.value = State.ShowHistory(withContext(Dispatchers.IO) { evaluationRecordRepository.getEvaluationHistory() })
+                }
             }
             is State.ShowHistory -> {
                 _state.value = State.ShowExpression(expression)
