+++ calculator/src/main/res/layout/activity_calculator.xml
@@ -8,7 +8,7 @@
         <variable
             name="viewModel"
             type="edu.nextstep.camp.calculator.CalculatorViewModel" />
-        <import type="edu.nextstep.camp.calculator.domain.Operator" />
+        <import type="edu.nextstep.camp.calculator.domain.model.Operator" />
     </data>
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
